Create Exceptions

+++ src/main/java/com/mathsena/timetodo/repository/TodoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TodoRepository extends JpaRepository<Todo, Long> {
+public interface TodoRepository extends JpaRepository<Todo, Integer> {
     @Query("SELECT obj FROM Todo obj WHERE obj.isFinished = false ORDER BY obj.timeToFinish")
     List<Todo> findAllOpen();
